+++ src/Base/Utilities.ts
@@ -1,10 +1,11 @@
+
 // https://en.wikipedia.org/wiki/Speed_reading
 const AVERAGE_READING_WORDS_PER_MINUTE = 200;
 const AVERAGE_READING_WORDS_PER_SECOND = AVERAGE_READING_WORDS_PER_MINUTE / 60;
 
 export const capitalize = (value: string) => `${value[0].toUpperCase()}${value.slice(1)}`;
 
-export const determineTimeToRead = (text: string | String, min: number = 3000): number => {
+export const timeToRead = (text: string | String, min: number = 3000): number => {
   const isString = (typeof text === 'string') || (text instanceof String);
   if (!isString) {
     return min;