+++ src/Store/Epic/NotificationEpic.ts
@@ -6,8 +6,6 @@ import { isActionOf, PayloadAction } from 'typesafe-actions';
 import {
   addNotificationAction,
   addNotificationWithIdAction,
-  dismissAllNotificationsAction,
-  dismissNextNotificationAction,
   fadeNotificationOffScreenAction,
   hideNotificationDrawerAction,
   pinNotificationDrawerAction,
@@ -17,7 +15,7 @@ import {
 } from '../Action/NotificationAction';
 import { disableSplitViewAction, enableSplitViewAction } from '../Action/SplitViewAction';
 import { INotification, INotificationContent } from '../Reducer/NotificationReducer';
-import { determineTimeToRead } from '../../Base/Utilities';
+import { timeToRead } from '../../Component/Notification/NotificationCard';
 
 let notificationIncrementId = 0;
 
@@ -31,7 +29,7 @@ export const addNotificationEpic: Epic<RootAction, RootAction> = (action$) =>
           id: ++notificationIncrementId,
           count: 1,
           timestamp: new Date(),
-          autoHideDelay: !notification.stayOnScreen ? determineTimeToRead(notification.message) : null
+          autoHideDelay: !notification.stayOnScreen ? timeToRead(notification) : null
         })
       ))
     ),
@@ -43,15 +41,6 @@ export const addNotificationEpic: Epic<RootAction, RootAction> = (action$) =>
           : EMPTY
       )
     ),
-  );
-
-export const dismissAllNotificationsEpic: Epic<RootAction, RootAction> = (action$, store) =>
-  action$.pipe(
-    filter(isActionOf([dismissAllNotificationsAction, dismissNextNotificationAction])),
-    mergeMap(() => store.value.notificationReducer.notifications.length
-      ? of(dismissNextNotificationAction()).pipe(delay(10))
-      : EMPTY
-    )
   );
 
 export const splitViewEpic: Epic<RootAction, RootAction> = (action$, store) =>