+++ CHANGELOG.md
@@ -1,6 +1,51 @@
+## [8.6.1] - 2020-03-17
+### Bugfix
+- Prevent websocket error to be raised more then 1 time during connection/reconnection
+- Fixed a bug that changeLanguageAction prop in SettingsContainer is not required anymore
+- Fixed a shadowed name issue in src/Store/Reducer/index.ts
+
+## [8.6.0] - 2020-03-17
+### Added
+- Full support of react-router functionalities for AuthRoute, NoAuthRoute & BaseRoute
+
+## [8.5.7] - 2020-03-16
+### Bugfixes
+- Fixed error handling in websocket context
+
+### Added
+- new config value for websocket connection: Path
+  * is optional: default is `/`
+
+## [8.5.6] - 2020-03-13
+### Bugfixes
+- Dismiss all Notifications will not dismiss those with dismissButtonVisible === false
+
+## [8.5.5] - 2020-03-13
+### Bugfixes
+- Smoother Notification animations
+
+## [8.5.4] - 2020-03-13
+### Added
+- Hide button to on screen notifications
+- Hide/Dismiss-Animation optimization
+
+## [8.5.3] - 2020-03-10
+### Bugfixes
+- invalid import shorthand
+
+## [8.5.2] - 2020-03-10
+### Added
+- Notification
+    - dismissButtonVisible option
+    - message type supports string[] to render as list
+    - translationId type supports string[] to render as list
+### Bugfixes
+- Notification
+    - timeToRead will also get calculated for translations and error objects
+
 ## [8.5.1] - 2020-03-10
 ### Bugfixes
-- fixed an issue when try to emit to websocket when no websocket connection is available 
+- fixed an issue when try to emit to websocket when no websocket connection is available
 
 ## [8.5.0] - 2020-03-10
 ### Added